gh actions: fix the build process again
On Windows we are building for both Debug & Release (CRT requirements). We
need to build the "default" configuration (Release) first, because the
tools such as moc & rcc from a non-default configuration will not be placed
in the right directory.
Previously we'd used a different solution (really a series of hacks) which
caused the mac arm64-on-x64 build to fail. This should be a clean & working
solution.